% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alzheimer-desc.R
\docType{data}
\name{AADR-alzheimer}
\alias{AADR-alzheimer}
\title{Number of deaths for alzheimers in the US from 1999-2013}
\format{Comma Separate Value file with 765 rows and 5 variables.}
\usage{
data(AADR-alzheimer)
}
\description{
A dataset containing the Age Adjusted Death Rates and number of 
deaths for alzheimers listed by region and year. The variables
are as follows:
}
\details{
\itemize{
 \item YEAR. The year in which the deaths occurred
 \item CAUSE_NAME. The cause of deaths
 \item region. The state in which the deaths occurred
 \item DEATHS. The number of Deaths via state and year
 \item AADR. Age Adjusted Death Rates via state and year
}
}
\keyword{datasets}

