% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accidents.R
\name{Accidents}
\alias{Accidents}
\title{Accidents}
\usage{
Accidents(x)
}
\arguments{
\item{x}{user inputed year from 1999-2013}
}
\value{
Choropleth of Accident related deaths
}
\description{
This function creates a choropleth of Accident related deaths by year.
}

