% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alzheimer.R
\name{alzheimer}
\alias{alzheimer}
\title{alzheimer}
\usage{
alzheimer(x)
}
\arguments{
\item{x}{user inputed year from 1999-2013}
}
\value{
Choropleth of Alzheimer related deaths
}
\description{
This function creates a choropleth of Alzheimer related deaths by year.
}

