% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancer.R
\name{cancer}
\alias{cancer}
\title{cancer}
\usage{
cancer(x)
}
\arguments{
\item{x}{user inputed year from 1999-2013}
}
\value{
Choropleth of cancer related deaths
}
\description{
This function creates a choropleth of Cancer related deaths by year.
}

