% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top10.R
\name{top10}
\alias{top10}
\title{top10}
\usage{
top10(x)
}
\arguments{
\item{x}{user inputed year from 1999-2013}
}
\value{
Colorpleth of the top 10 types of death
}
\description{
This function creates a colorpleth of the Top 10 deaths by year.
}

